Support dtype

+++ tvmffi_test.py
@@ -3,7 +3,7 @@ import time
 from mxnet.ndarray import np
 
 print("tvm ffi...")
-a = np.zeros1((3, 4), ctx="cpu(0)")
+a = np.zeros1((3, 4), ctx="cpu(0)", dtype='float64')
 print(a)
 print("legacy ffi...")
 a = np.zeros((3, 4), ctx=mx.cpu())
@@ -13,7 +13,7 @@ print("tvm ffi dummy...")
 repeat = 10000
 start = time.time()
 for i in range(repeat):
-    a = np.zeros0((3, 4), ctx="cpu(0)")
+    a = np.zeros0((3, 4), ctx="cpu(0)", dtype='float64')
 end = time.time()
 print("time = {}".format((end - start) / repeat))
 
@@ -21,6 +21,6 @@ print("tvm ffi...")
 repeat = 10000
 start = time.time()
 for i in range(repeat):
-    a = np.zeros1((3, 4), ctx="cpu(0)")
+    a = np.zeros1((3, 4), ctx="cpu(0)", dtype='float64')
 end = time.time()
 print("time = {}".format((end - start) / repeat))
